+++ android/src/main/java/com/swmansion/rnscreens/bottomsheet/SheetDelegate.kt
@@ -1,6 +1,12 @@
 package com.swmansion.rnscreens.bottomsheet
 
+import android.content.Context
+import android.os.Build
+import android.util.Log
 import android.view.View
+import android.view.WindowManager
+import android.view.inputmethod.InputMethodManager
+import androidx.annotation.RequiresApi
 import androidx.core.graphics.Insets
 import androidx.core.view.OnApplyWindowInsetsListener
 import androidx.core.view.WindowInsetsCompat
@@ -8,6 +14,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.LifecycleOwner
 import com.google.android.material.bottomsheet.BottomSheetBehavior
+import com.google.android.material.bottomsheet.BottomSheetBehavior.BottomSheetCallback
 import com.swmansion.rnscreens.InsetsObserverProxy
 import com.swmansion.rnscreens.KeyboardDidHide
 import com.swmansion.rnscreens.KeyboardNotVisible
@@ -23,13 +30,16 @@ class SheetDelegate(
 
     private var isKeyboardVisible: Boolean = false
     private var keyboardState: KeyboardState = KeyboardNotVisible
-    private var lastStableDetentIndex: Int = screen.sheetInitialDetentIndex
+
+    var lastStableDetentIndex: Int = screen.sheetInitialDetentIndex
+        private set
 
     @BottomSheetBehavior.State
-    private var lastStableState: Int = SheetUtils.sheetStateFromDetentIndex(
+    var lastStableState: Int = SheetUtils.sheetStateFromDetentIndex(
         screen.sheetInitialDetentIndex,
         screen.sheetDetents.count()
     )
+        private set
 
     private val sheetStateObserver = SheetStateObserver()
 
@@ -85,12 +95,160 @@ class SheetDelegate(
                 newState,
                 screen.sheetDetents.count()
             )
+
+            Log.i(
+                TAG,
+                "lastStableState: $lastStableState, lastStableDetentIndex: $lastStableDetentIndex"
+            )
         }
 
         screen.onSheetDetentChanged(lastStableDetentIndex, isStable)
 
         if (shouldDismissSheetInState(newState)) {
             stackFragment.dismissSelf()
+        }
+    }
+
+    fun onSheetBehaviorAttached(behavior: BottomSheetBehavior<Screen>) {
+        // There is a guard internally that does not allow the callback to be duplicated.
+        behavior.addBottomSheetCallback(sheetStateObserver)
+    }
+
+    internal fun configureBottomSheetBehaviour(
+        behavior: BottomSheetBehavior<Screen>,
+        keyboardState: KeyboardState = KeyboardNotVisible,
+        selectedDetentIndex: Int = lastStableDetentIndex,
+    ): BottomSheetBehavior<Screen> {
+
+        val containerHeight = tryResolveContainerHeight()
+        check(containerHeight != null) {
+            "[RNScreens] Failed to find window height during bottom sheet behaviour configuration"
+        }
+
+        behavior.apply {
+            isHideable = true
+            isDraggable = true
+        }
+
+        Log.i(TAG, "Configure with selectedDetentIndex: $selectedDetentIndex")
+
+        screen.footer?.registerWithSheetBehavior(behavior)
+
+        return when (keyboardState) {
+            is KeyboardNotVisible -> {
+                when (screen.sheetDetents.count()) {
+                    1 ->
+                        behavior.apply {
+                            val height =
+                                if (screen.isSheetFitToContents()) {
+                                    screen.contentWrapper
+                                        .get()
+                                        ?.height
+                                        .takeIf { screen.contentWrapper.get()?.isLaidOut == true }
+                                } else {
+                                    (screen.sheetDetents.first() * containerHeight).toInt()
+                                }
+                            useSingleDetent(height = height)
+                        }
+
+                    2 ->
+                        behavior.useTwoDetents(
+                            state =
+                            SheetUtils.sheetStateFromDetentIndex(
+                                selectedDetentIndex,
+                                screen.sheetDetents.count(),
+                            ),
+                            firstHeight = (screen.sheetDetents[0] * containerHeight).toInt(),
+                            secondHeight = (screen.sheetDetents[1] * containerHeight).toInt(),
+                        )
+
+                    3 ->
+                        behavior.useThreeDetents(
+                            state =
+                            SheetUtils.sheetStateFromDetentIndex(
+                                selectedDetentIndex,
+                                screen.sheetDetents.count(),
+                            ),
+                            firstHeight = (screen.sheetDetents[0] * containerHeight).toInt(),
+                            halfExpandedRatio = (screen.sheetDetents[1] / screen.sheetDetents[2]).toFloat(),
+                            expandedOffsetFromTop = ((1 - screen.sheetDetents[2]) * containerHeight).toInt(),
+                        )
+
+                    else -> throw IllegalStateException(
+                        "[RNScreens] Invalid detent count ${screen.sheetDetents.count()}. Expected at most 3.",
+                    )
+                }
+            }
+
+            is KeyboardVisible -> {
+                val newMaxHeight =
+                    if (behavior.maxHeight - keyboardState.height > 1) {
+                        behavior.maxHeight - keyboardState.height
+                    } else {
+                        behavior.maxHeight
+                    }
+                when (screen.sheetDetents.count()) {
+                    1 ->
+                        behavior.apply {
+                            useSingleDetent(height = newMaxHeight)
+                            addBottomSheetCallback(keyboardSheetCallback)
+                        }
+
+                    2 ->
+                        behavior.apply {
+                            useTwoDetents(
+                                state = BottomSheetBehavior.STATE_EXPANDED,
+                                secondHeight = newMaxHeight,
+                            )
+                            addBottomSheetCallback(keyboardSheetCallback)
+                        }
+
+                    3 ->
+                        behavior.apply {
+                            useThreeDetents(
+                                state = BottomSheetBehavior.STATE_EXPANDED,
+                            )
+                            maxHeight = newMaxHeight
+                            addBottomSheetCallback(keyboardSheetCallback)
+                        }
+
+                    else -> throw IllegalStateException(
+                        "[RNScreens] Invalid detent count ${screen.sheetDetents.count()}. Expected at most 3.",
+                    )
+                }
+            }
+
+            is KeyboardDidHide -> {
+                // Here we assume that the keyboard was either closed explicitly by user,
+                // or the user dragged the sheet down. In any case the state should
+                // stay unchanged.
+
+                behavior.removeBottomSheetCallback(keyboardSheetCallback)
+                when (screen.sheetDetents.count()) {
+                    1 ->
+                        behavior.useSingleDetent(
+                            height = (screen.sheetDetents.first() * containerHeight).toInt(),
+                            forceExpandedState = false,
+                        )
+
+                    2 ->
+                        behavior.useTwoDetents(
+                            firstHeight = (screen.sheetDetents[0] * containerHeight).toInt(),
+                            secondHeight = (screen.sheetDetents[1] * containerHeight).toInt(),
+                        )
+
+                    3 ->
+                        behavior.useThreeDetents(
+                            firstHeight = (screen.sheetDetents[0] * containerHeight).toInt(),
+                            halfExpandedRatio = (screen.sheetDetents[1] / screen.sheetDetents[2]).toFloat(),
+                            expandedOffsetFromTop = ((1 - screen.sheetDetents[2]) * containerHeight).toInt(),
+                        )
+
+                    else -> throw IllegalStateException(
+                        "[RNScreens] Invalid detent count ${screen.sheetDetents.count()}. Expected at most 3.",
+                    )
+                }
+            }
         }
     }
 
@@ -106,7 +264,7 @@ class SheetDelegate(
             isKeyboardVisible = true
             keyboardState = KeyboardVisible(imeInset.bottom)
             sheetBehavior?.let {
-                stackFragment.configureBottomSheetBehaviour(it, keyboardState)
+                this.configureBottomSheetBehaviour(it, keyboardState)
             }
 
             val prevInsets = insets.getInsets(WindowInsetsCompat.Type.navigationBars())
@@ -124,9 +282,9 @@ class SheetDelegate(
         } else {
             sheetBehavior?.let {
                 if (isKeyboardVisible) {
-                    stackFragment.configureBottomSheetBehaviour(it, KeyboardDidHide)
+                    this.configureBottomSheetBehaviour(it, KeyboardDidHide)
                 } else if (keyboardState != KeyboardNotVisible) {
-                    stackFragment.configureBottomSheetBehaviour(it, KeyboardNotVisible)
+                    this.configureBottomSheetBehaviour(it, KeyboardNotVisible)
                 } else {
                 }
             }
@@ -146,6 +304,67 @@ class SheetDelegate(
 
     private fun shouldDismissSheetInState(@BottomSheetBehavior.State state: Int) =
         state == BottomSheetBehavior.STATE_HIDDEN
+
+    /**
+     * This method might return slightly different values depending on code path,
+     * but during testing I've found this effect negligible. For practical purposes
+     * this is acceptable.
+     */
+    private fun tryResolveContainerHeight(): Int? {
+        screen.container?.let { return it.height }
+
+        val context = screen.reactContext
+
+        context
+            .resources
+            ?.displayMetrics
+            ?.heightPixels
+            ?.let { return it }
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+            (context.getSystemService(Context.WINDOW_SERVICE) as? WindowManager)
+                ?.currentWindowMetrics
+                ?.bounds
+                ?.height()
+                ?.let { return it }
+        }
+        return null
+    }
+
+    private val keyboardSheetCallback =
+        object : BottomSheetCallback() {
+            @RequiresApi(Build.VERSION_CODES.M)
+            override fun onStateChanged(
+                bottomSheet: View,
+                newState: Int,
+            ) {
+                if (newState == BottomSheetBehavior.STATE_COLLAPSED) {
+                    val isImeVisible =
+                        WindowInsetsCompat
+                            .toWindowInsetsCompat(bottomSheet.rootWindowInsets)
+                            .isVisible(WindowInsetsCompat.Type.ime())
+                    if (isImeVisible) {
+                        // Does it not interfere with React Native focus mechanism? In any case I'm not aware
+                        // of different way of hiding the keyboard.
+                        // https://stackoverflow.com/questions/1109022/how-can-i-close-hide-the-android-soft-keyboard-programmatically
+                        // https://developer.android.com/develop/ui/views/touch-and-input/keyboard-input/visibility
+
+                        // I want to be polite here and request focus before dismissing the keyboard,
+                        // however even if it fails I want to try to hide the keyboard. This sometimes works...
+                        bottomSheet.requestFocus()
+                        val imm =
+                            screen.reactContext.getSystemService(InputMethodManager::class.java)
+                        imm.hideSoftInputFromWindow(bottomSheet.windowToken, 0)
+                    }
+                }
+            }
+
+            override fun onSlide(
+                bottomSheet: View,
+                slideOffset: Float,
+            ) = Unit
+        }
+
 
     private inner class SheetStateObserver : BottomSheetBehavior.BottomSheetCallback() {
         override fun onStateChanged(bottomSheet: View, newState: Int) {
